Added: Jumping VFX
Added the jumping animation of the player. This is done with the animator and connecting the Player_Jump with the Player_Fall and the Player_Idle with eachother. Then I created a bool that holds the condition of what state it's going into and change it in code in teh PlayerController.